Move registry into contrib
Registry is a runtime dependency of the k8s broker - a store
of definitions of service classes available via k8s broker.
