+++ contrib/registry/Dockerfile
