+++ contrib/registry/README.md
