+++ contrib/registry/data/charts/README.md
