+++ contrib/registry/data/charts/booksbe/.helmignore
