+++ contrib/registry/data/charts/booksbe/Chart.yaml
