+++ contrib/registry/data/charts/booksbe/schema.yaml
