+++ contrib/registry/data/charts/booksbe/templates/NOTES.txt
