+++ contrib/registry/data/charts/booksbe/templates/config.yaml
