+++ contrib/registry/data/charts/booksbe/templates/sql.yaml
