+++ contrib/registry/data/charts/booksbe/templates/svc.yaml
