+++ contrib/registry/data/charts/booksbe/values.yaml
