+++ contrib/registry/data/charts/booksfe/.helmignore
