+++ contrib/registry/data/charts/booksfe/Chart.yaml
