+++ contrib/registry/data/charts/booksfe/schema.yaml
