+++ contrib/registry/data/charts/booksfe/templates/_helpers.tpl
