+++ contrib/registry/data/charts/booksfe/templates/nodejs.yaml
