+++ contrib/registry/data/charts/booksfe/templates/svc.yaml
