+++ contrib/registry/data/charts/booksfe/values.yaml
