+++ contrib/registry/data/charts/definitions.json
