+++ contrib/registry/data/charts/update.sh
