+++ contrib/registry/data/charts/user-provided-client/Chart.yaml
