+++ contrib/registry/data/charts/user-provided-client/schema.yaml
