+++ contrib/registry/data/charts/user-provided-client/templates/configmap.yaml
