+++ contrib/registry/data/charts/user-provided-client/values.yaml
