+++ contrib/registry/data/types/bq-dataset/dataset-binding.jinja.schema
