+++ contrib/registry/data/types/bq-dataset/dataset.jinja.schema
