+++ contrib/registry/data/types/test/config.yaml
