+++ contrib/registry/data/types/test/test-binding.py.schema
