+++ contrib/registry/registry_test.go
