+++ contrib/registry/server/controller.go
