+++ contrib/registry/server/server.go
