+++ contrib/registry/server/server_suite_test.go
